# AMDGPU: add parallel restore of BO content to accelerate restore