--- a/plugins/amdgpu/amdgpu_socket_utils.h
+++ b/plugins/amdgpu/amdgpu_socket_utils.h
@@ -1,0 +1,6 @@
+#ifndef __KFD_PLUGIN_AMDGPU_SOCKET_UTILS_H__
+#define __KFD_PLUGIN_AMDGPU_SOCKET_UTILS_H__
+
+int install_parallel_sock(void);
+
+#endif