@rst0git 2024-11-25
@Ddnirvana  @wweewrwer Thank you for contributions! It might be good to also update the content of the following file to reflect these changes: https://github.com/checkpoint-restore/criu/blob/criu-dev/plugins/amdgpu/README.md