@Ddnirvana 2024-11-25
> @Ddnirvana @wweewrwer Thank you for your contributions! It might be good to also update the content of the following files to reflect these changes:
> 
> * https://github.com/checkpoint-restore/criu/blob/criu-dev/plugins/amdgpu/README.md
> * https://github.com/checkpoint-restore/criu/blob/criu-dev/Documentation/criu-amdgpu-plugin.txt

@rst0git No problem. We will add proper description in the next version.