@wweewrwer 2024-11-28
@rst0git @avagin @dayatsin-amd hi maintainers, thanks for your prior reviews and comments. We have fixed all the issues, as the following:
1. Use the proper APIs to allocate (xmalloc, etc.)
2. Enable the optimizations by default
3. Change the name of hook
4. Fix the issues to run in Podman containers
5. Other fixes (line width, comments, etc.) 
6. Add descriptions in README to explain the optimizations.

Please let us know if you have any further comments