@rst0git 2024-11-28
@wweewrwer Would you be able to merge the fixup commits into the previous commits using `git rebase`?
https://github.com/checkpoint-restore/criu/blob/criu-dev/CONTRIBUTING.md#submit-your-work-upstream