@wweewrwer 2024-11-29
> @wweewrwer Would you be able to merge the fixup commits into the previous commits using `git rebase`? https://github.com/checkpoint-restore/criu/blob/criu-dev/CONTRIBUTING.md#submit-your-work-upstream

Thanks for your comment! I have merged the fixup commits into the previous commits using git rebase. Please let me know if you have any further comments.