@wweewrwer 2024-12-19
> I have requested this PR to be validated on a multi-GPU set-up internally at AMD. Can you give us a few days to confirm there is no regression.
> 
> Thank you for this patch!

Sure. Thanks!