@avagin 2025-01-07
> > Have you investigated other approaches of parallel restoring of BO? For example, it is possible to fork a thread and restoring BO asynchronously in context of its process. In this case, two BO will be restored concurrently.
> 
> Yes, we have investigated the approach of forking a thread in the background, but it cannot work as it conflicts with the restore logic of CRIU.
> 
> Specifically, when CRIU tries to restore its memory state, it will unmap all old mappings. However, some mappings may be needed by the background thread for BO restoring. Therefore, a thread can only run in parallel with shorter procedures (possibly before entering the restorer blob), while offloading the restore of BO content to a new process (in this PR) can be parallelized with almost the entire restore procedure.

Everything what is happening in the restore blob should be fast, the main restore process is happening before switching to the restorer blob. I am still not convinced that we need to offload that to another process... Maybe I miss something, but I want to see numbers and clear explanation why the current approach is good.

This new approach solves one parallelism issue by introducing another one. In this new version, all buffer objects are restored consequently. It means that it can be even slower for workloads with many buffer objects (BO).

Another problem here is that a plugin callback is running through the entire restore process